final integration tests, prep for cold flow

+++ FireController/src/StateClock/StateClock.h
@@ -10,6 +10,13 @@ public:
     {
         _total_clock_start = micros();
         _state_clock_start = micros();
+        // _total_clock = 0;
+        // _state_clock = 0;
+        // _total_dt = 0;
+        // _state_dt = 0;
+        // _total_ticks = 0;
+        // _state_ticks = 0;
+        // _should_update_state = false;
     }
 
     void tick()
@@ -40,22 +47,22 @@ public:
         _should_update_state = false;
     }
 
-    long total_et()
+    long long total_et()
     {
         return _total_clock;
     }
 
-    long total_et_ms()
+    long long total_et_ms()
     {
         return _total_clock / 1000;
     }
 
-    long state_et()
+    long long state_et()
     {
         return _state_clock;
     }
 
-    long state_et_ms()
+    long long state_et_ms()
     {
         return _state_clock / 1000;
     }
